pyproject.toml
[tool.poetry]
name = "backup-to-gcp"
version = "0.1.0"
description = "Simple utility for backup to Google Cloud Storage"
authors = ["Saeed Gholami Shahbandi <saeed.gh.sh@gmail.com>"]
license = "GNU GENERAL PUBLIC"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
tqdm = "^4.66.1"
google-cloud-storage = "^2.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
black = "^23.10.0"
pylint = "^3.0.1"
mypy = "^1.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-p no:warnings"
